Address review feedback: fix lazy logging, IntegrityError handling, test organization
- Convert remaining f-string logger in export_v2.py to lazy % formatting
- Wrap get_or_create in _import_ingestion_sources with IntegrityError handler
  for concurrent import safety
- Move out-of-scope TestDRFMutationValidationError from test_ingestion_source.py
  to test_security_hardening.py where DRFMutation tests belong
- Extract duplicate _upload_mutation methods into module-level
  UPLOAD_DOCUMENT_MUTATION constant per DRY principle
